+++ GUI/algebra_ui.py
@@ -1,6 +1,6 @@
-from slope1 import *
-from slope2 import *
-from pythagore import *
+from slope1_ui import *
+from slope2_ui import *
+from pythagore_ui import *
 from PyQt5 import QtCore, QtGui, QtWidgets
