+++ GUI/arithmetics_ui.py
