+++ GUI/pythagore_ui.py
