+++ GUI/slope1_ui.py
