+++ GUI/slope2_ui.py
