+++ GUI/stock_ui.py
